Discord/requirements.txt
blurple.py==0.5.5
discord.py==1.7.3
python-dotenv==0.20.0
t3sf==1.0